portfolio/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.9.1
brotli==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cssselect2==0.8.0
distro==1.9.0
Django==5.2.5
fonttools==4.61.1
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.3.0
pycparser==2.23
pydantic==2.13.3
pydantic_core==2.46.3
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
